+++ src/PickerPanel/TimePanel/index.tsx
@@ -2,8 +2,8 @@ import classNames from 'classnames';
 import * as React from 'react';
 import { formatValue } from '../../utils/dateUtil';
 import type { SharedPanelProps } from '../../interface';
-import { PanelContext, useInfo } from '../../PickerPanel/context';
-import PanelHeader from '../../PickerPanel/PanelHeader';
+import { PanelContext, useInfo } from '../context';
+import PanelHeader from '../PanelHeader';
 import TimePanelBody from './TimePanelBody';
 
 export type TimePanelProps<DateType extends object> = SharedPanelProps<DateType>;
